+++ packages/reactabular-easy/CHANGELOG.md
@@ -3,6 +3,7 @@
 
   * Improvement - Add sorting numbers to the header so you know in which order sorting rules are applied.
   * Improvement - Allow column sorting status to be reset by doubleclicking on a column header.
+  * Improvement - Add support for row selection. Capture the selected row through `onSelectRow`.
 
 1.0.8 / 2016-07-28
 ==================